extract PackageSwiftItemParser for reuse and scale
