+++ Sources/PackageSwiftLSPLibrary/PackageSwift/PackageSwiftItemLocator.swift
@@ -3,10 +3,12 @@ import SwiftSyntax
 final class PackageSwiftItemLocator {
     let tree: SourceFileSyntax
     let locationConverter: SourceLocationConverter
+    let parser: PackageSwiftItemParser
 
     init(tree: SourceFileSyntax, locationConverter: SourceLocationConverter) {
         self.tree = tree
         self.locationConverter = locationConverter
+        self.parser = PackageSwiftItemParser.defaultParser(locationConverter: locationConverter)
     }
 
     func item(at position: OneBasedPosition) -> PackageSwiftItem? {
@@ -19,71 +21,15 @@ final class PackageSwiftItemLocator {
             return nil
         }
 
-        return findCompletionContext(from: node, position: position)
-    }
-
-    private func findCompletionContext(from node: Syntax, position: AbsolutePosition) -> PackageSwiftItem? {
-        var current: Syntax? = node
-
+        var current: SyntaxProtocol? = node
         while let node = current {
-            if let call = node.as(FunctionCallExprSyntax.self),
-               let member = call.calledExpression.as(MemberAccessExprSyntax.self),
-               let functionKind = PackageSwiftItem.FunctionKind.from(memberName: member.declName.baseName.text),
-               let arguments = PackageSwiftItem.NonEmptyFunctionArguments(
-                   arguments: extractArguments(from: call, for: functionKind),
-                   activeArgumentIndex: findCurrentArgumentIndex(in: call, at: position)
-               )
-            {
-                switch functionKind {
-                case .package:
-                    return .packageFunctionCall(arguments: arguments)
-                case .product:
-                    return .productFunctionCall(arguments: arguments)
-                }
+            if let parsedItem = parser.parse(node, position) {
+                return parsedItem
             }
 
             current = node.parent
         }
 
         return nil
-    }
-
-    private func findCurrentArgumentIndex(
-        in call: FunctionCallExprSyntax,
-        at position: AbsolutePosition
-    ) -> Int? {
-        for (index, argument) in call.arguments.enumerated() {
-            let start = argument.positionAfterSkippingLeadingTrivia
-            let end = argument.endPositionBeforeTrailingTrivia
-            if start <= position, position <= end {
-                return index
-            }
-        }
-        return nil
-    }
-
-    private func extractArguments(
-        from call: FunctionCallExprSyntax,
-        for function: PackageSwiftItem.FunctionKind
-    ) -> [PackageSwiftItem.FunctionArgument] {
-        let allowed = PackageSwiftItem.FunctionArgumentLabel.allowedLabels(for: function)
-
-        return call.arguments.compactMap { argument -> PackageSwiftItem.FunctionArgument? in
-            guard let labelText = argument.label?.text,
-                  let label = PackageSwiftItem.FunctionArgumentLabel.from(label: labelText),
-                  let stringLiteralExpression = argument.expression.as(StringLiteralExprSyntax.self)?.segments,
-                  allowed.contains(label)
-            else {
-                return nil
-            }
-
-            let value = stringLiteralExpression.trimmedDescription
-
-            let range = stringLiteralExpression.trimmedRange
-            let startPosition = OneBasedPosition(locationConverter.location(for: range.lowerBound))
-            let endPosition = OneBasedPosition(locationConverter.location(for: range.upperBound))
-
-            return .init(label: label, stringValueRange: .init(start: startPosition, end: endPosition), stringValue: value)
-        }
     }
 }
